Project Functional
ConcurrentModificationException was being thrown due to a call to remove an arraylist item within a for loop utilizing the same arraylist. I moved the call out of the for loop and used a boolean flag to trigger the removal after the for loop call.

No exceptions are thrown now.

Thermometer value grows as soon as a donor is added to the display queue, as opposed to when the donor starts to be displayed. Fixing is non-essential.

Thermometer design could use work, but is also non-essential.

Project is functional and could be used without modification.